+++ src/renderer/editors/base/EditorError.tsx
@@ -1,0 +1,11 @@
+import styled from "@emotion/styled";
+import color from "../../theme/color";
+
+const EditorError = styled.div({
+    whiteSpace: "pre-wrap",
+    margin: "auto",
+    padding: 24,
+    color: color.misc.yellow,
+});
+
+export { EditorError };